[REFACTOR]: menuId, menuFolderId를 Int -> Long으로 수정

+++ app/src/main/java/com/kuit/ourmenu/data/repository/MenuFolderRepository.kt
@@ -14,7 +14,7 @@ class MenuFolderRepository @Inject constructor(
     }
 
     suspend fun getMenuFolderDetail(
-        menuFolderId: Int,
+        menuFolderId: Long,
         sortOrder: String,
     ) = runCatching {
         menuFolderService.getMenuFolderDetails(
